split the car index into separate files
