+++ Cargo.toml
@@ -147,6 +147,7 @@ sha2 = { version = "0.10.5", default-features = false }
 shared_memory = "0.12"
 similar = "2.2.1"
 slotmap = "1.0"
+smallvec = "1.11"
 smart-default = "0.7.1"
 static_assertions = "1.1.0"
 statrs = "0.16"
